requirements.txt
asgiref==3.5.2
coverage==6.5.0
Django==4.1.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
django-filter==22.1
Faker==15.3.2
faker_music==0.4
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3